fixed read in of floating values for inspecticion rate

+++ dft2lnt/dft_ast.h
@@ -465,7 +465,7 @@ public:
     }
     
     int getPhases() const {return phases;}
-    int getLambda() const {return lambda;}
+    double getLambda() const {return lambda;}
 };
 
 /**
